# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mindful AI Sandbox API
  version: 1.0.0
  description: |
    Mindful AI Sandbox is a student learning environment for building and experimenting
    with cloud-native applications on AWS.

    **Key Features:**
    - User authentication and role-based access control (User, Admin)
    - User profile management
    - Admin user management

    **Authentication:**
    Most endpoints require JWT authentication via the `Authorization` header.
    Use the `/auth/login` endpoint to obtain an access token.

  contact:
    name: Mindful AI Sandbox Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/dev
    description: Development server
  - url: https://api.example.com/staging
    description: Staging server
  - url: https://api.example.com/prod
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /me:
    get:
      tags: [Profile]
      summary: Get current user profile
      description: Retrieve authenticated user's private profile (includes email).
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Profile]
      summary: Update current user profile
      description: Update profile information (username, bio, avatar).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              username: student42
              bio: Learning cloud-native development
              avatarUrl: https://example.com/avatar.jpg
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{userId}:
    get:
      tags: [Profile]
      summary: Get public user profile
      description: Retrieve public profile for any user (excludes email).
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
          example: user-123
      responses:
        '200':
          description: Public profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfile'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================================
  # ADMIN USER MANAGEMENT ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  schemas:
...
    UserRole:
      type: string
      enum: [User, Admin]
      description: User role in the system
...
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_-]+$'
        bio:
          type: string
          maxLength: 500
        avatarUrl:
          type: string
          format: uri
    PrivateProfile:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: string
                  example: "38510330-9041-70d5-010c-a30b00bd812d"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                username:
                  type: string
                  example: "test_user"
                role:
                  $ref: '#/components/schemas/UserRole'
                bio:
                  type: string
                  nullable: true
                  example: "Learning cloud-native development"
                avatarUrl:
                  type: string
                  nullable: true
                  example: "https://example.com/avatar.jpg"
                createdAt:
                  type: string
                  format: date-time
                  example: "2025-10-28T06:35:13.807Z"
                updatedAt:
                  type: string
                  format: date-time
                  example: "2025-11-01T21:03:23.065Z"
    PublicProfile:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: string
                  example: "38510330-9041-70d5-010c-a30b00bd812d"
                username:
                  type: string
                  example: "test_user"
                role:
                  $ref: '#/components/schemas/UserRole'
                bio:
                  type: string
                  nullable: true
                  example: "Learning cloud-native development"
                avatarUrl:
                  type: string
                  nullable: true
                  example: "https://example.com/avatar.jpg"
                createdAt:
                  type: string
                  format: date-time
                  example: "2025-10-28T06:35:13.807Z"
    UpdateProfileResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponseWrapper'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
                  example: "Profile updated successfully"
                profile:
                  type: object
                  description: Updated profile data (structure same as PrivateProfile data field)

    # =========================================================================
    # ADMIN USER MANAGEMENT SCHEMAS
    # =========================================================================
...
    ResponseMeta:
      type: object
      required: [timestamp, requestId]
      properties:
        timestamp:
          type: string
          format: date-time
          description: Server timestamp when response was generated
          example: "2025-11-01T21:03:23.065Z"
        requestId:
          type: string
          format: uuid
          description: Unique request ID for debugging and tracing
          example: "ca1fec38-e588-4be7-acbc-921575189fbe"
    SuccessResponseWrapper:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          description: Indicates successful response
          example: true
        data:
          description: Response payload (structure varies by endpoint, can be object or array)
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorResponseWrapper:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          description: Indicates failed response
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for client-side error handling
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input data"
            details:
              type: object
              nullable: true
              description: Additional error details (field-level errors, etc.)
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
  responses:
    ValidationError:
      description: Validation error (invalid request data)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                validationErrors:
                  - field: email
                    message: Invalid email format
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
    UnauthorizedError:
      description: Unauthorized (missing or invalid authentication)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
    ConflictError:
      description: Conflict (resource already exists or state conflict)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: CONFLICT
              message: Resource already exists
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWrapper'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
            meta:
              timestamp: "2025-11-01T21:03:23.065Z"
              requestId: "ca1fec38-e588-4be7-acbc-921575189fbe"